Feature: add user common types to pb impl, add unit test cases
