+++ common/proto-conv/src/lib.rs
@@ -19,8 +19,9 @@
 //! This mod will convert protobuf message to the current version of meta data used in databend-query.
 
 mod data_from_to_protobuf_impl;
+mod database_from_to_protobuf_impl;
 mod from_to_protobuf;
-mod from_to_protobuf_impl;
+mod table_from_to_protobuf_impl;
 mod user_from_to_protobuf_impl;
 mod util;
 
@@ -28,3 +29,4 @@ pub use from_to_protobuf::FromToProto;
 pub use from_to_protobuf::Incompatible;
 pub use util::check_ver;
 pub use util::missing;
+pub use util::VER;
