+++ common/proto-conv/tests/it/main.rs
@@ -12,3 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 mod proto_conv;
+mod user_proto_conv;
